requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.3.7
aiohttp==3.10.4
aiolimiter==1.1.0
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
backoff==2.2.1
bcrypt==4.2.0
black==25.1.0
build==1.2.1
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.22
click==8.1.7
cloudpickle==3.1.1
coloredlogs==15.0.1
contourpy==1.2.1
cramjam==2.10.0
cvxopt==1.3.2
cycler==0.12.1
dataclasses-json==0.6.7
Deprecated==1.2.14
deprecation==2.1.0
distro==1.9.0
etils==1.5.2
exceptiongroup==1.2.2
fastapi==0.112.1
fastcore==1.8.5
fastparquet==2024.11.0
filelock==3.15.4
filetype==1.2.0
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
google-ai-generativelanguage==0.6.18
google-api-core==2.25.1
google-auth==2.33.0
googleapis-common-protos==1.63.2
graphviz==0.20.3
greenlet==3.2.3
grpcio==1.65.5
grpcio-status==1.71.2
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.24.5
humanfriendly==10.0
humanize==4.12.3
ibm-generative-ai==3.0.0
idna==3.7
importlib_metadata==8.0.0
importlib_resources==6.4.3
intervaltree==3.1.0
jax==0.4.30
jaxlib==0.4.30
jaxtyping==0.2.36
jiter==0.5.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.5
kubernetes==30.1.0
langchain==0.2.14
langchain-community==0.2.12
langchain-core==0.2.33
langchain-google-genai==2.1.6
langchain-openai==0.1.22
langchain-text-splitters==0.2.2
langgraph==0.5.1
langgraph-checkpoint==2.1.0
langgraph-prebuilt==0.5.2
langgraph-sdk==0.1.72
langsmith==0.1.99
lingam==1.9.0
llvmlite==0.43.0
lxml==5.3.0
markdown-it-py==3.0.0
marshmallow==3.21.3
matplotlib==3.9.2
mdurl==0.1.2
ml_dtypes==0.5.1
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
msgpack==1.1.1
multidict==6.0.5
mypy-extensions==1.0.0
natsort==8.4.0
nest-asyncio==1.6.0
networkx==3.2.1
numba==0.60.0
numdifftools==0.9.41
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.19.0
openai==1.41.0
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-instrumentation==0.47b0
opentelemetry-instrumentation-asgi==0.47b0
opentelemetry-instrumentation-fastapi==0.47b0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
opentelemetry-util-http==0.47b0
opt_einsum==3.4.0
orjson==3.10.7
ormsgpack==1.10.0
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pathspec==0.12.1
patsy==0.5.6
pillow==10.4.0
platformdirs==4.3.8
pm4py==2.7.11.12
posthog==3.5.0
progressbar2==4.4.2
proto-plus==1.26.1
protobuf==4.25.4
psy==0.0.1
pulsar-client==3.7.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pybind11==2.13.6
pydantic==2.8.2
pydantic_core==2.20.1
pydotplus==2.0.2
pygam==0.9.1
Pygments==2.18.0
pyparsing==3.1.2
pypdf==4.3.1
PyPika==0.48.9
pyproject_hooks==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-utils==3.8.2
pytz==2024.1
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.7.1
rsa==4.9
sax4bpm==0.3.1
scikit-learn==1.5.1
scipy==1.11.4
semopy==2.3.11
shap==0.48.0
shellingham==1.5.4
six==1.16.0
slicer==0.0.8
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.32
starlette==0.38.2
statsmodels==0.14.2
sympy==1.13.2
tenacity==8.5.0
tensorstore==0.1.69
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.20.0
tomli==2.0.1
toolz==1.0.0
tqdm==4.66.5
typer==0.12.4
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.6
uvloop==0.20.0
watchfiles==0.23.0
websocket-client==1.8.0
websockets==12.0
wrapt==1.16.0
xgboost==2.1.4
xxhash==3.5.0
yarl==1.9.4
zipp==3.20.0
zstandard==0.23.0
